progress - task (a) finished
The task for the first part of the assignment was to create a node that
implements an action client, allowing the user to set the target(x,y) or
to cancel it. Additionally there was a suggestion to take advantage of
feedback and status of the action server to know when the target has
been reached. The node also should also publsih robots position and
velocity in a form of a custom message (x,y, vel_x, vel_y) by relying on
the values published on the topic /odom.

This approach to the above problem is purely based on callbacks to
specific events.

To set the new target/goal for the action server, the user should publish
a custom message onto `/robot_target` topic with x and y coordinates:
eg. `rostopic pub /robot_target rt_assignment_2/robotTarget "{x: 1.0, y: 1.0}"`
When new target arrives, the previous one is preempted and robot will
focus on the new one.

In order to cancel the goal, the user should publis a custom message
onto `/robot_cancel_goal` topic eg:
`rostopic pub /robot_cancel_goal rt_assignment_2/robotCancelGoal "{cancelGoal: true}"`

The node has 3 subscribers. First two are meant for above operations.
The last one subscribes to `/odom` to get the position and velocity in
both x and y, and then publish them onto `/robot_state` topic.

Additionally to that, there is a timer that is logging the current
robot's position from action server's feedback when the current goal
state will be active.