Task (b) done
The second task was to implement a service node that when called, should
return the last target sent by the user.

The task was accomplished by creating a an object of service server that
has a subscriber to `/robot_target` topic, where user is publishing the
new target msg. Then service, when called is returning the target
position unless it was not yet published.